+++ DMSport/Sources/App/SceneDelegate.swift
@@ -12,7 +12,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
                 window = UIWindow(frame: UIScreen.main.bounds)
                 window?.windowScene = windowScene
-                window?.rootViewController = MainLoginViewController()
+                window?.rootViewController = SignUpViewController()
                 window?.makeKeyAndVisible()
     }
